refactor: clean up imports and improve error handling in API interceptors
